requirements.txt
datasets==2.16.0
faster_rcnn.egg==info
matplotlib==3.8.2
numpy==1.26.4
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.1.0
Pillow==10.2.0
pycocotools==2.0.7
scipy==1.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
